pyproject.toml
[project]
name = "do-android-wolves-dream-of-electric-sheep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.61.1",
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]
